refactor verify handler
