+++ auth-service/routes/auth.test.ts
@@ -53,154 +53,6 @@ describe("Auth Routes", () => {
     vi.clearAllMocks();
   });
 
-  describe("GET /auth/verify", () => {
-    it("should return 401 when not authenticated", async () => {
-      const response = await request(app).get("/auth/verify");
-
-      expect(response.status).toBe(401);
-      expect(response.body).toEqual({ message: "Unauthorized!" });
-    });
-
-    it("should return user ID and email when authenticated", async () => {
-      // Setup test user
-      const userData = {
-        email: "test@example.com",
-        password: "password123",
-      };
-
-      const user = {
-        id: 1,
-        email: userData.email,
-        createdAt: new Date(),
-        lastLoginAt: null,
-      };
-
-      // Mock database responses
-      vi.spyOn(db.users, "getByEmail").mockResolvedValue(user);
-      vi.spyOn(db.emailAuth, "getByEmail").mockResolvedValue({
-        userId: user.id,
-        email: user.email,
-        passwordHash: Buffer.from("hashed-password"),
-        verifiedAt: null,
-        verificationToken: null,
-        verificationTokenExpiresAt: null,
-        forgotPasswordToken: null,
-        forgotPasswordTokenExpiresAt: null,
-      });
-      vi.spyOn(db.users, "getById").mockResolvedValue(user);
-      vi.spyOn(db.users, "updateLastLogin").mockResolvedValue({
-        ...user,
-        lastLoginAt: new Date(),
-      });
-      vi.spyOn(argon2, "verify").mockResolvedValue(true);
-
-      // Use agent to maintain cookies between requests
-      const agent = request.agent(app);
-
-      // Login first to establish session
-      const loginResponse = await agent.post("/auth/login").send(userData);
-
-      expect(loginResponse.status).toBe(200);
-
-      // Then verify authentication
-      const verifyResponse = await agent.get("/auth/verify");
-
-      expect(verifyResponse.status).toBe(200);
-      expect(verifyResponse.body).toEqual({
-        id: user.id,
-        email: user.email,
-        scopes: ["none"],
-      });
-      expect(verifyResponse.header["x-user-id"]).toBe(user.id.toString());
-      expect(verifyResponse.header["x-user-email"]).toBe(user.email);
-      expect(verifyResponse.header["x-user-scopes"]).toBe("none");
-    });
-
-    it("should handle health check requests", async () => {
-      const response = await request(app)
-        .get("/auth/verify")
-        .set("x-forwarded-uri", "/health");
-
-      expect(response.status).toBe(200);
-      expect(response.body).toEqual({});
-    });
-
-    it("should handle OPTIONS requests", async () => {
-      const response = await request(app)
-        .get("/auth/verify")
-        .set("x-forwarded-method", "OPTIONS");
-
-      expect(response.status).toBe(200);
-      expect(response.body).toEqual({});
-    });
-
-    it("should return user scopes when user has admin permission", async () => {
-      // Setup test user
-      const userData = {
-        email: "admin@example.com",
-        password: "password123",
-      };
-
-      const user = {
-        id: 1,
-        email: userData.email,
-        createdAt: new Date(),
-        lastLoginAt: null,
-      };
-
-      // Mock database responses
-      vi.spyOn(db.users, "getByEmail").mockResolvedValue(user);
-      vi.spyOn(db.emailAuth, "getByEmail").mockResolvedValue({
-        userId: user.id,
-        email: user.email,
-        passwordHash: Buffer.from("hashed-password"),
-        verifiedAt: null,
-        verificationToken: null,
-        verificationTokenExpiresAt: null,
-        forgotPasswordToken: null,
-        forgotPasswordTokenExpiresAt: null,
-      });
-      vi.spyOn(db.users, "getById").mockResolvedValue(user);
-      vi.spyOn(db.users, "updateLastLogin").mockResolvedValue({
-        ...user,
-        lastLoginAt: new Date(),
-      });
-      vi.spyOn(argon2, "verify").mockResolvedValue(true);
-
-      // Mock permissions to return admin scope
-      vi.spyOn(db.permissions, "getByUserId").mockResolvedValue([
-        {
-          id: 1,
-          userId: user.id,
-          permission: "admin",
-          createdAt: new Date(),
-          grantedBy: user.id,
-        },
-      ]);
-
-      // Use agent to maintain cookies between requests
-      const agent = request.agent(app);
-
-      // Login first to establish session
-      const loginResponse = await agent.post("/auth/login").send(userData);
-
-      expect(loginResponse.status).toBe(200);
-
-      // Then verify authentication
-      const verifyResponse = await agent.get("/auth/verify");
-
-      expect(verifyResponse.status).toBe(200);
-      expect(verifyResponse.body).toEqual({
-        id: user.id,
-        email: user.email,
-        scopes: ["admin"],
-      });
-      expect(verifyResponse.header["x-user-id"]).toBe(user.id.toString());
-      expect(verifyResponse.header["x-user-email"]).toBe(user.email);
-      expect(verifyResponse.header["x-user-scopes"]).toBe("admin");
-    });
-  });
-
   describe("POST /auth/forgot-password", () => {
     it("should generate and store reset token when user exists", async () => {
       const email = "test@example.com";
